Do not implement unlawful Ord for LockedDependency
